feat: migrate comments api

+++ src/routes/school/[schoolId]/comments/+page.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import { PUBLIC_API_HOST } from '$env/static/public';
 	import { PUBLIC_NOCO_TOKEN_KEY } from '$env/static/public';
 
 	import { onMount } from 'svelte';
@@ -24,39 +25,23 @@
 	let txt_comment = '';
 	let chk_locations: string[] = [];
 
+	$: api_query = `locations=${encodeURIComponent(
+		filter_locations.join(',')
+	)}&years=${encodeURIComponent(filter_years.join(','))}&sort=${filter_sort_by}`;
+
+	const assignCommentsVariables = async (resp: Response) => {
+		const json = await resp.json();
+		posts = json?.list.map((e: any) => ({
+			...e,
+			likedByYourself: e.likedUserIds.find((e: any) => e.userId === $currentUser?.uid)
+		}));
+	};
+
 	const fetchComments = async () => {
-		let schoolid_query = `schoolId,eq,${schoolId}`;
-		let approved_query = `approved,eq,true`;
-		let location_query = filter_locations.length
-			? filter_locations.map((loc) => `(location,like,${loc})`).join('~or')
-			: null;
-		let year_query = filter_years.length
-			? filter_years.map((year) => `(schoolYear,eq,${year})`).join('~or')
-			: null;
-		let where_query = [schoolid_query, approved_query, location_query, year_query]
-			.filter((e) => e)
-			.map((e) => `(${e})`)
-			.join('~and');
-
-		let sort_query = filter_sort_by === 'latest' ? '-createDate' : '-likeCount';
-
 		try {
-			const resp = await fetch(
-				`https://sheets.wevis.info/api/v1/db/data/v1/Open-School-Test/SchoolComments?limit=999&where=${encodeURIComponent(
-					where_query
-				)}&sort=${encodeURIComponent(sort_query)}`,
-				{
-					method: 'GET',
-					headers: {
-						'xc-token': PUBLIC_NOCO_TOKEN_KEY
-					}
-				}
-			);
-			const json = await resp.json();
-			posts = json?.list.map((e: any) => ({
-				...e,
-				likedByYourself: e.likedUserIds.find((e: any) => e.userId === $currentUser?.uid)
-			}));
+			const resp = await fetch(`${PUBLIC_API_HOST}/schools/${schoolId}/comments?${api_query}`);
+
+			await assignCommentsVariables(resp);
 		} catch (err) {
 			console.error(err);
 		}
@@ -66,20 +51,25 @@
 		if (!$currentUser) return;
 
 		try {
-			await fetch(
-				`https://sheets.wevis.info/api/v1/db/data/v1/Open-School-Test/SchoolCommentLike`,
+			const resp = await fetch(
+				`${PUBLIC_API_HOST}/schools/${schoolId}/comments/${commentId}/like`,
 				{
 					method: 'POST',
 					headers: {
-						'xc-token': PUBLIC_NOCO_TOKEN_KEY,
+						Authorization: `Bearer ${$currentUser.accessToken}`,
 						'Content-Type': 'application/json'
 					},
 					body: JSON.stringify({
-						userId: $currentUser.uid,
-						commentId: commentId
+						schoolCommentsQuery: {
+							locations: filter_locations,
+							years: filter_years,
+							sort: filter_sort_by
+						}
 					})
 				}
 			);
+
+			await assignCommentsVariables(resp);
 		} catch (err) {
 			console.error(err);
 		} finally {
@@ -88,45 +78,62 @@
 	};
 
 	const unlikeComment = async (likeId: number) => {
+		if (!$currentUser) return;
+
 		try {
-			await fetch(
-				`https://sheets.wevis.info/api/v1/db/data/v1/Open-School-Test/SchoolCommentLike/${likeId}`,
-				{
-					method: 'DELETE',
-					headers: {
-						'xc-token': PUBLIC_NOCO_TOKEN_KEY
+			const resp = await fetch(`${PUBLIC_API_HOST}/schools/${schoolId}/likes/${likeId}`, {
+				method: 'DELETE',
+				headers: {
+					Authorization: `Bearer ${$currentUser.accessToken}`,
+					'Content-Type': 'application/json'
+				},
+				body: JSON.stringify({
+					schoolCommentsQuery: {
+						locations: filter_locations,
+						years: filter_years,
+						sort: filter_sort_by
 					}
-				}
-			);
+				})
+			});
+
+			await assignCommentsVariables(resp);
 		} catch (err) {
 			console.error(err);
-		} finally {
-			fetchComments();
 		}
 	};
 
 	const deleteComment = async () => {
-		const commentId = going_to_delete_id;
+		if (!$currentUser) return;
+
 		try {
-			await fetch(
-				`https://sheets.wevis.info/api/v1/db/data/v1/Open-School-Test/SchoolComments/${commentId}`,
+			const resp = await fetch(
+				`${PUBLIC_API_HOST}/schools/${schoolId}/comments/${going_to_delete_id}`,
 				{
 					method: 'DELETE',
 					headers: {
-						'xc-token': PUBLIC_NOCO_TOKEN_KEY
-					}
+						Authorization: `Bearer ${$currentUser.accessToken}`,
+						'Content-Type': 'application/json'
+					},
+					body: JSON.stringify({
+						schoolCommentsQuery: {
+							locations: filter_locations,
+							years: filter_years,
+							sort: filter_sort_by
+						}
+					})
 				}
 			);
+
+			await assignCommentsVariables(resp);
 		} catch (err) {
 			console.error(err);
 		} finally {
 			going_to_delete_id = null;
 			confirm_delete_isopen = false;
-			fetchComments();
 		}
 	};
 
-	const post = async () => {
+	const addComment = async () => {
 		if (!$currentUser) return;
 		if (txt_comment.trim() === '' || chk_locations.length === 0) return;
 
@@ -151,25 +158,27 @@
 			uploaded_files = await resp.json();
 		}
 
-		// console.log(uploaded_files);
-
 		try {
-			await fetch('https://sheets.wevis.info/api/v1/db/data/v1/Open-School-Test/SchoolComments', {
-				method: 'POST',
+			const resp = await fetch(`${PUBLIC_API_HOST}/schools/${schoolId}/comments`, {
+				method: 'PUT',
 				headers: {
-					'xc-token': PUBLIC_NOCO_TOKEN_KEY,
+					Authorization: `Bearer ${$currentUser.accessToken}`,
 					'Content-Type': 'application/json'
 				},
 				body: JSON.stringify({
-					schoolId: schoolId,
 					comments: txt_comment.trim(),
 					location: chk_locations.join(),
 					schoolYear: $LATEST_YEAR,
-					userId: $currentUser.uid,
-					approved: true || !uploaded_files,
-					...(uploaded_files ? { images: JSON.stringify(uploaded_files) } : null)
+					...(uploaded_files ? { images: JSON.stringify(uploaded_files) } : null),
+					schoolCommentsQuery: {
+						locations: filter_locations,
+						years: filter_years,
+						sort: filter_sort_by
+					}
 				})
 			});
+
+			await assignCommentsVariables(resp);
 		} catch (err) {
 			console.error(err);
 		} finally {
@@ -340,7 +349,7 @@
 			class="cf-submit"
 			type="button"
 			disabled={txt_comment.trim() === '' || chk_locations.length === 0}
-			on:click={post}
+			on:click={addComment}
 			slot="title"
 		>
 			ส่งความเห็น
